+++ nkc_modules/api_operations/gallery.js
@@ -1,12 +1,12 @@
 var table = {};
 module.exports = table;
 
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
+
 var moment = require('moment') //packages you may need
 var fs = require('fs.extra')
-var settings = require('server_settings.js');
-var helper_mod = require('helper.js')();
-var queryfunc = require('query_functions')
+var settings = require('../server_settings.js');
+var helper_mod = require('../helper.js')();
+var queryfunc = require('../query_functions')
 var AQL = queryfunc.AQL
 
 var layer = require('../layer')
