+++ nkc_modules/api_operations/mail.js
@@ -1,23 +1,23 @@
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
+
 
 var moment = require('moment')
 var path = require('path')
 var fs = require('fs.extra')
-var settings = require('server_settings.js');
-var helper_mod = require('helper.js')();
-var queryfunc = require('query_functions')
-var validation = require('validation')
+var settings = require('../server_settings.js');
+var helper_mod = require('../helper.js')();
+var queryfunc = require('../query_functions')
+var validation = require('../validation')
 var AQL = queryfunc.AQL
-var apifunc = require('api_functions')
-var layer = require('../layer')
+var apifunc = require('../api_functions')
+var layer = require('../layer');
 
-var nm = require('nodemailer')
+var nm = require('nodemailer');
 
 try{
-  var mailSecrets = require('mailSecrets.js')
+  var mailSecrets = require('../mailSecrets.js')
 }
 catch(e){
-  var mailSecrets = require('mailSecrets_template.js')
+  var mailSecrets = require('../mailSecrets_template.js')
 }
 
 var transporter = nm.createTransport(mailSecrets.smtpConfig);
