+++ nkc_modules/api_operations/management.js
@@ -1,11 +1,11 @@
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
+
 
 var moment = require('moment')
 var path = require('path')
 var fs = require('fs.extra')
-var settings = require('server_settings.js');
-var helper_mod = require('helper.js')();
-var queryfunc = require('query_functions')
+var settings = require('../server_settings.js');
+var helper_mod = require('../helper.js')();
+var queryfunc = require('../query_functions')
 var AQL = queryfunc.AQL
 
 var layer = require('../layer')
@@ -85,7 +85,7 @@ table.disablePost = {
     .then(()=>{
       return post.update({disabled:true})
       .then(post=>{
-        var op = require('api_operations')
+        var op = require('../api_operations.js')
         return op.table.updateThread.operation({tid:post.model.tid})
         .then(()=>{
           return 'success'
@@ -368,7 +368,7 @@ table.addCredit = {
         toc:Date.now(),
       })
       .then(cl=>{
-        var operations = require('api_operations')
+        var operations = require('../api_operations.js')
         return operations.table.updatePost.operation(params)
       })
       .then(()=>{
