+++ nkc_modules/api_operations/search.js
@@ -1,11 +1,11 @@
 var table = {};
 module.exports = table;
 
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
+
 var moment = require('moment') //packages you may need
-var settings = require('server_settings.js');
-var helper_mod = require('helper.js')();
-var queryfunc = require('query_functions')
+var settings = require('../server_settings.js');
+var helper_mod = require('../helper.js')();
+var queryfunc = require('../query_functions')
 var AQL = queryfunc.AQL
 
 //this is an example API
