+++ nkc_modules/jade/interface_profile_info.jade
@@ -8,7 +8,10 @@
         //.ForumTitleCircle(style='background-color:#{user.color||"#bbb"}')
         .ForumName #{user.username}&nbsp;&nbsp;&nbsp;
           if(thatuser._key != data.user._key)
-            a.btn.btn-sm.btn-info(href:'javascript: subscribeUserSwitch(#{uid})')#subscribeButton 订阅
+            if(data.user.subscribeUser.indexOf(thatuser._key) > -1)
+              a.btn.btn-sm.btn-danger(href='javascript: unsubscribeUser(#{uid});')#subscribeButton 取关
+            else
+              a.btn.btn-sm.btn-info(href='javascript: subscribeUser(#{uid});')#subscribeButton 关注
     -var user = thatuser
     include interface_profile_info_section2.jade
