+++ nkc_modules/permissions.js
@@ -1,6 +1,6 @@
-module.paths.push('./nkc_modules'); //enable require-ment for this path
 
-var operations = require('api_operations');
+
+var operations = require('./api_operations');
 var table = operations.table;
 
 var permissions = {};
