+++ nkc_modules/promisify.js
@@ -1,4 +1,4 @@
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
+
 
 function promisify(func){
   return function(){
