+++ nkc_modules/validation.js
@@ -1,5 +1,5 @@
 //validation module
-module.paths.push('./nkc_modules'); //enable require-ment for this path
+
 
 //decide whether a submitted post is legal
 exports.validatePost = function(p){
