+++ taravaEater.js
@@ -1,11 +1,9 @@
 global.__projectroot = __dirname + '/';//create global variable for project root directory
-module.paths.push(__projectroot + 'nkc_modules'); //enable require-ment for this path
-module.paths.push(__projectroot ); //enable require-ment for this path
 
 var fs = require('fs-extra')
 var dir = require('node-dir');
 
-var fpath = require('tEaterSettings.js').avatar_root
+var fpath = require('./tEaterSettings.js').avatar_root
 
 var dpath = __projectroot + 'resources/newavatar'
 var spath = __projectroot + 'resources/newavatar_small'
